// src/types/reducer-type.ts
import { CartItems, ShippingInfo, User } from "./types";

export interface UserReducerInitialState {
    user: User | null;
    loading: boolean;
}
export interface CartReducerInitialState {
    loading: boolean;
    cartItems: CartItems[];
    subtotal: number;
    tax: number;
    shippingCharges: number;
    discount: number;
    total: number;
    shippingInfo: ShippingInfo;
}